phan quyen cho phan SYS

+++ trunk/Sources/Source_Codes/FBDSource/FBD/Controllers/SYSBranchesController.cs
@@ -20,6 +20,10 @@ namespace FBD.Controllers
         /// <returns>[Index] view</returns>
         public ActionResult Index()
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_VIEW, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             List<SystemBranches> branches = null;
 
             try
@@ -48,6 +52,10 @@ namespace FBD.Controllers
         /// <returns>[Add] view</returns>
         public ActionResult Add()
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             return View();
         }
 
@@ -67,6 +75,10 @@ namespace FBD.Controllers
         [HttpPost]
         public ActionResult Add(SystemBranches branch)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 if (ModelState.IsValid)
@@ -113,6 +125,10 @@ namespace FBD.Controllers
         /// [Index] view: if ERROR</returns>
         public ActionResult Edit(string id)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             SystemBranches branch = null;
 
             try
@@ -150,6 +166,10 @@ namespace FBD.Controllers
         [HttpPost]
         public ActionResult Edit(string id, SystemBranches branch)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 if (ModelState.IsValid)
@@ -186,6 +206,10 @@ namespace FBD.Controllers
         /// <returns>[Index] view</returns>
         public ActionResult Delete(string id)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_SYSTEM_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 int result = SystemBranches.DeleteBranch(id);
